backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
altair==5.5.0
anyio==4.10.0
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==2.1.1
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dash==3.2.0
dask==2025.7.0
dnspython==2.3.0
fastapi==0.100.0
filelock==3.24.3
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
fonttools==4.61.0
frozenlist==1.7.0
fsspec==2025.9.0
geoip2==5.1.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
gunicorn==20.1.0
h11==0.16.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
locket==1.0.0
loguru==0.7.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.7
maxminddb==2.8.2
motor==3.7.1
multidict==6.6.4
narwhals==2.1.1
nest-asyncio==1.6.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
partd==1.4.2
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
plotly==6.3.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.1.0
pyarrow==21.0.0
pycparser==2.23
pydantic==1.10.12
pydeck==0.9.1
PyJWT==2.6.0
pymongo==4.14.1
pynput==1.8.1
pyparsing==3.1.1
pypdfium2==5.0.0
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-engineio==4.12.3
python-multipart==0.0.6
python-socketio==5.15.0
python-whois==0.9.6
pytz==2025.2
pywin32==311
PyYAML==6.0.2
RapidFuzz==3.14.3
referencing==0.36.2
requests==2.28.1
requests-file==3.0.1
retrying==1.4.2
rpds-py==0.27.0
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
shiboken6==6.10.1
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.27.0
streamlit==1.48.1
tenacity==9.1.2
threadpoolctl==3.6.0
tldextract==5.3.1
toml==0.10.2
toolz==1.0.0
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.22.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
win32_setctime==1.2.0
wsproto==1.3.2
yarl==1.20.1
zipp==3.23.0
